Add export as a function
