+++ .gitignore
@@ -81,3 +81,4 @@ lib
 
 # next.js build output
 .next
+.idea
