+++ src/run.ts
@@ -1,0 +1,4 @@
+#! /usr/bin/env node
+import { command } from './command';
+
+command();
